fix the integration test
