+++ test/integration/operator/agent/mgh/suite_test.go
@@ -1,0 +1,80 @@
+package mgh
+
+import (
+	"context"
+	"os"
+	"path/filepath"
+	"testing"
+	"time"
+
+	_ "github.com/lib/pq"
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+	"k8s.io/client-go/rest"
+	ctrl "sigs.k8s.io/controller-runtime"
+	"sigs.k8s.io/controller-runtime/pkg/envtest"
+	logf "sigs.k8s.io/controller-runtime/pkg/log"
+	"sigs.k8s.io/controller-runtime/pkg/log/zap"
+	"sigs.k8s.io/controller-runtime/pkg/manager"
+	metricsserver "sigs.k8s.io/controller-runtime/pkg/metrics/server"
+
+	"github.com/stolostron/multicluster-global-hub/operator/pkg/config"
+)
+
+var (
+	testenv *envtest.Environment
+	cfg     *rest.Config
+	ctx     context.Context
+	cancel  context.CancelFunc
+	mgr     manager.Manager
+)
+
+func TestController(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Addons Controller Suite")
+}
+
+var _ = BeforeSuite(func() {
+	logf.SetLogger(zap.New(zap.WriteTo(GinkgoWriter), zap.UseDevMode(true)))
+	Expect(os.Setenv("POD_NAMESPACE", "default")).To(Succeed())
+
+	ctx, cancel = context.WithCancel(context.Background())
+
+	By("Prepare envtest environment")
+	var err error
+	testenv = &envtest.Environment{
+		CRDDirectoryPaths: []string{
+			filepath.Join("..", "..", "..", "..", "manifest", "crd"),
+			filepath.Join("..", "..", "..", "..", "..", "operator", "config", "crd", "bases"),
+		},
+		ErrorIfCRDPathMissing: true,
+	}
+	cfg, err = testenv.Start()
+	Expect(err).NotTo(HaveOccurred())
+	Expect(cfg).NotTo(BeNil())
+	mgr, err = ctrl.NewManager(cfg, ctrl.Options{
+		Metrics: metricsserver.Options{
+			BindAddress: "0", // disable the metrics serving
+		},
+		Scheme: config.GetRuntimeScheme(),
+	})
+	Expect(err).NotTo(HaveOccurred())
+	go func() {
+		Expect(mgr.Start(ctx)).NotTo(HaveOccurred())
+	}()
+
+	By("Waiting for the manager to be ready")
+	Expect(mgr.GetCache().WaitForCacheSync(ctx)).To(BeTrue())
+})
+
+var _ = AfterSuite(func() {
+	By("tearing down the test environment")
+	err := testenv.Stop()
+	// https://github.com/kubernetes-sigs/controller-runtime/issues/1571
+	// Set 4 with random
+	if err != nil {
+		time.Sleep(4 * time.Second)
+		Expect(testenv.Stop()).NotTo(HaveOccurred())
+	}
+	cancel()
+})
